Reformulated some packages and code generation centralized
