+++ python/forsyde/io/python/core.py
@@ -1,11 +1,18 @@
-from dataclasses import dataclass, field
-from typing import Optional, Set, Any, Dict, Tuple
+from dataclasses import dataclass
+from dataclasses import field
+from typing import Iterable
+from typing import Dict
+from typing import Any
+from typing import Optional
+from typing import Set
+from typing import Tuple
+
 
 _port_id_counter = 1
 _vertex_id_counter = 1
 
 
-class Type(object):
+class ModelType(object):
     """Type associated with a vertex or a port.
     
     Though Python already keeps many runtime amenities that would make this
@@ -15,20 +22,38 @@ class Type(object):
 
     This clas is meant to be used more of a interface than a concrete class.
     """
+
+    _instance = None
+
+    @classmethod
+    def get_instance(cls):
+        if not cls._instance:
+            cls._instance = cls()
+        return cls._instance
+
+    def __repr__(self):
+        return self.get_type_name()
+
+    def __eq__(self, other: "ModelType"):
+        return self.get_type_name() == other.get_type_name()
+
     def __hash__(self):
         return hash(self.get_type_name())
 
-    def is_refinement(self, other: "Type") -> bool:
-        return False
+    def is_refinement(self, other: "ModelType") -> bool:
+        return other == self or any(s.is_refinement(other) for s in self.get_super_types())
+
+    def get_super_types(self) -> Iterable["ModelType"]:
+        yield self
 
     def get_type_name(self) -> str:
-        return "Unknown"
+        return "UnknownType"
 
-    def get_required_ports(self) -> Dict[str, Optional[str]]:
-        return dict()
+    def get_required_ports(self) -> Iterable[Tuple[str, "ModelType"]]:
+        yield from ()
 
-    def get_required_properties(self) -> Dict[str, Any]:
-        return dict()
+    def get_required_properties(self) -> Iterable[Tuple[str, Any]]:
+        yield from ()
 
 
 @dataclass
@@ -42,7 +67,7 @@ class Port(object):
     """
 
     identifier: Optional[str] = field(default=None, hash=True)
-    port_type: Type = field(default=Type(), compare=False, hash=False)
+    port_type: ModelType = field(default=ModelType(), compare=False, hash=False)
 
     def __post_init__(self):
         global _port_id_counter
@@ -53,22 +78,27 @@ class Port(object):
     def __hash__(self):
         return hash(self.identifier)
 
-    def is_type(self, t: Type) -> bool:
+    def is_type(self, t: ModelType) -> bool:
         return self.port_type.is_refinement(t)
 
 
 @dataclass
 class Vertex(object):
-    """Docstring for Vertex. """
+    """Class holding data regarding Vertexes.
+
+    Every vertex representes a main element in a ForSyDe model.
+    Every vertex contains a number of ports (which are repeated in the
+    vertexed to increase reliability in the model, since putting
+    them in edges would have been sufficient) with their associated types.
+    Also, every port contains "Properties" which are arbitrary associated data,
+    such as the size of bits in a Signal or the time slots in a Time Division
+    Multiplexer.
+    """
 
     identifier: Optional[str] = field(default=None, hash=True)
-    ports: Set[Port] = field(default_factory=lambda: set(),
-                             compare=False,
-                             hash=False)
-    properties: Dict[str, Any] = field(default_factory=lambda: dict(),
-                                       compare=False,
-                                       hash=False)
-    vertex_type: Type = field(default=Type(), compare=False, hash=False)
+    ports: Set[Port] = field(default_factory=lambda: set(), compare=False, hash=False)
+    properties: Dict[str, Any] = field(default_factory=lambda: dict(), compare=False, hash=False)
+    vertex_type: ModelType = field(default=ModelType(), compare=False, hash=False)
 
     def __post_init__(self):
         global _vertex_id_counter
@@ -79,29 +109,32 @@ class Vertex(object):
     def __hash__(self):
         return hash(self.identifier)
 
-    def is_type(self, t: Type) -> bool:
+    def is_type(self, t: ModelType) -> bool:
         return self.vertex_type.is_refinement(t)
 
 
 @dataclass
 class Edge(object):
-    """Docstring for Edge. """
+    """Class containing all information for an Edge.
+
+    The edge contains references to the source and target vetexes
+    as well as the ports being connect on both ends, in case
+    they exist. The edges also have types associated with them
+    so that extra deductions can be made along the EDA flow.
+    """
     source_vertex: Vertex
     target_vertex: Vertex
     source_vertex_port: Optional[Port] = field(default=None)
     target_vertex_port: Optional[Port] = field(default=None)
-    edge_type: Type = field(default=Type(), compare=False)
+    edge_type: ModelType = field(default=ModelType(), compare=False)
 
     def __hash__(self):
         return hash((self.source_vertex, self.target_vertex))
 
     def ids_tuple(self):
         return (self.source_vertex.identifier, self.target_vertex.identifier,
-                self.source_vertex_port.identifier if self.source_vertex_port
-                else None, self.target_vertex_port.identifier
-                if self.target_vertex_port else None,
-                self.edge_type.get_type_name())
+                self.source_vertex_port.identifier if self.source_vertex_port else None,
+                self.target_vertex_port.identifier if self.target_vertex_port else None, self.edge_type.get_type_name())
 
-    def is_type(self, tsource: Type, ttarget: Type) -> bool:
-        return self.source_vertex.is_type(
-            tsource) and self.target_vertex.is_type(ttarget)
+    def is_type(self, tsource: ModelType, ttarget: ModelType) -> bool:
+        return self.source_vertex.is_type(tsource) and self.target_vertex.is_type(ttarget)
